refactor: remove mix files and rename config
